Fix for the MongoDB issue where the collection was wiped out sometimes
The problem was likely that the unit tests ran deleteAll() on the
repository. I'm not sure how to switch collections and still test the
methods in the repository. In the tests, I made a special user name
and just work with that, deleting all docs indexed using it. This seems
to fix the problem. There is a way to use MongoTemplate to switch
collections, but I'm not sure how to use that and then test the methods
built by spring
boot in the FeedbackRepository.